refactor: 소셜 토큰 검증 로직을 OAuthAttributesResolver로 추출
