+++ app-main/src/main/java/net/causw/app/main/domain/user/account/service/SocialLinkService.java
@@ -16,8 +16,8 @@ import net.causw.app.main.domain.user.account.service.implementation.SocialAccou
 import net.causw.app.main.domain.user.account.service.implementation.UserReader;
 import net.causw.app.main.domain.user.account.service.implementation.UserValidator;
 import net.causw.app.main.domain.user.account.service.implementation.UserWriter;
-import net.causw.app.main.domain.user.auth.service.SocialNativeAuthService;
 import net.causw.app.main.domain.user.auth.service.dto.OAuthAttributes;
+import net.causw.app.main.domain.user.auth.service.implementation.OAuthAttributesResolver;
 import net.causw.app.main.shared.exception.errorcode.AuthErrorCode;
 
 import lombok.RequiredArgsConstructor;
@@ -30,7 +30,7 @@ public class SocialLinkService {
 	private final UserWriter userWriter;
 	private final UserValidator userValidator;
 	private final SocialAccountReader socialAccountReader;
-	private final SocialNativeAuthService socialNativeAuthService;
+	private final OAuthAttributesResolver oAuthAttributesResolver;
 	private final SocialAccountLinker socialAccountLinker;
 
 	/**
@@ -65,7 +65,7 @@ public class SocialLinkService {
 	 */
 	@Transactional
 	public void linkSocialAccount(String userId, String provider, String accessToken, String idToken) {
-		OAuthAttributes attributes = socialNativeAuthService.extractAttributes(provider, accessToken, idToken);
+		OAuthAttributes attributes = oAuthAttributesResolver.resolveAttributes(provider, accessToken, idToken);
 		socialAccountLinker.applyLinkingPolicy(userId, attributes.socialType(), attributes.socialId(),
 			attributes.email());
 	}
