+++ app-main/src/main/java/net/causw/app/main/domain/user/auth/service/SocialNativeAuthService.java
@@ -1,20 +1,13 @@
 package net.causw.app.main.domain.user.auth.service;
 
-import java.time.Instant;
-import java.util.List;
 import java.util.Locale;
 import java.util.Optional;
 
 import org.springframework.security.authentication.InternalAuthenticationServiceException;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.OAuth2AccessToken;
-import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.security.oauth2.jwt.Jwt;
-import org.springframework.security.oauth2.jwt.JwtDecoder;
-import org.springframework.security.oauth2.jwt.JwtDecoderFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
@@ -26,8 +19,8 @@ import net.causw.app.main.domain.user.account.enums.user.SocialType;
 import net.causw.app.main.domain.user.auth.service.dto.AuthResult;
 import net.causw.app.main.domain.user.auth.service.dto.AuthTokenPair;
 import net.causw.app.main.domain.user.auth.service.dto.CustomOAuth2User;
-import net.causw.app.main.domain.user.auth.service.dto.OAuthAttributes;
 import net.causw.app.main.domain.user.auth.service.implementation.AuthTokenManager;
+import net.causw.app.main.domain.user.auth.service.implementation.OAuthAttributesResolver;
 import net.causw.app.main.domain.user.auth.service.implementation.OidcAuthorizationCodeTokenClient;
 import net.causw.app.main.domain.user.auth.service.implementation.SocialAccountOauthRefreshStore;
 import net.causw.app.main.domain.user.terms.service.implementation.UserTermsAgreementReader;
@@ -53,19 +46,14 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class SocialNativeAuthService {
 
-	private static final long PROVIDER_TOKEN_DEFAULT_TTL_SECONDS = 300L;
-	private static final String OIDC_SUB_CLAIM = "sub";
-	private static final String APPLE_ISSUER = "https://appleid.apple.com";
-	private static final String GOOGLE_ISSUER = "https://accounts.google.com";
-
 	private final ClientRegistrationRepository clientRegistrationRepository;
 	private final CustomOAuth2UserService customOAuth2UserService;
-	private final JwtDecoderFactory<ClientRegistration> oidcIdTokenDecoderFactory;
 	private final AuthTokenManager authTokenManager;
 	private final UserTermsAgreementReader userTermsAgreementReader;
 	private final UserProfileImageReader userProfileImageReader;
 	private final OidcAuthorizationCodeTokenClient oidcAuthorizationCodeTokenClient;
 	private final SocialAccountOauthRefreshStore socialAccountOauthRefreshStore;
+	private final OAuthAttributesResolver oAuthAttributesResolver;
 
 	/**
 	 * provider 특성에 따라 access token 또는 id token 기반 네이티브 소셜 로그인을 수행합니다.
@@ -94,7 +82,7 @@ public class SocialNativeAuthService {
 				.orElseThrow(AuthErrorCode.UNSUPPORTED_SOCIAL_PROVIDER::toBaseException);
 
 			User user = loadAuthenticatedUser(socialType, clientRegistration, accessToken, idToken);
-			if (isOidcProvider(clientRegistration)) {
+			if (oAuthAttributesResolver.isOidcProvider(clientRegistration)) {
 				ClientRegistration tokenExchangeRegistration = resolveOidcTokenExchangeRegistration(socialType,
 					clientRegistration);
 				persistOidcRefreshTokenFromAuthorizationCode(tokenExchangeRegistration, socialType, user,
@@ -122,46 +110,11 @@ public class SocialNativeAuthService {
 	}
 
 	/**
-	 * provider 토큰을 검증하고 소셜 계정 속성만 반환합니다. 계정 생성/연동은 수행하지 않습니다.
-	 * 마이페이지 소셜 계정 연동 시 토큰 유효성 검증 및 socialId/socialType 추출 용도로 사용합니다.
-	 *
-	 * @param provider    provider registration id (kakao, google, apple)
-	 * @param accessToken 카카오용 access token
-	 * @param idToken     구글/애플용 id token
-	 * @return 표준화된 소셜 계정 속성 (socialId, socialType, email 포함)
-	 */
-	public OAuthAttributes extractAttributes(String provider, String accessToken, String idToken) {
-		String registrationId = provider.toLowerCase(Locale.ROOT);
-		log.info("Social account link token extraction requested. provider={}", registrationId);
-
-		try {
-			SocialType.from(registrationId);
-
-			ClientRegistration clientRegistration = Optional.ofNullable(
-				clientRegistrationRepository.findByRegistrationId(registrationId))
-				.orElseThrow(AuthErrorCode.UNSUPPORTED_SOCIAL_PROVIDER::toBaseException);
-
-			if (isOidcProvider(clientRegistration)) {
-				return extractOidcAttributes(clientRegistration, registrationId, idToken);
-			}
-
-			return extractOAuth2Attributes(clientRegistration, registrationId, accessToken);
-		} catch (BaseRunTimeV2Exception e) {
-			log.warn("Social account link token extraction failed. provider={}, code={}, message={}",
-				registrationId, e.getErrorCode().getCode(), e.getMessage());
-			throw e;
-		} catch (RuntimeException e) {
-			log.error("Unexpected error during social account link token extraction. provider={}", registrationId, e);
-			throw e;
-		}
-	}
-
-	/**
 	 * provider 설정(openid scope 여부)에 따라 OAuth2(access token) 또는 OIDC(id token) 검증 경로를 선택합니다.
 	 */
 	private User loadAuthenticatedUser(SocialType socialType, ClientRegistration clientRegistration,
 		String providerAccessToken, String providerIdToken) {
-		if (isOidcProvider(clientRegistration)) {
+		if (oAuthAttributesResolver.isOidcProvider(clientRegistration)) {
 			return loadOidcAuthenticatedUser(socialType, clientRegistration, providerIdToken);
 		}
 
@@ -176,7 +129,7 @@ public class SocialNativeAuthService {
 			clientRegistration.getRegistrationId());
 
 		OAuth2UserRequest userRequest = new OAuth2UserRequest(clientRegistration,
-			toProviderAccessToken(providerAccessToken));
+			oAuthAttributesResolver.toProviderAccessToken(providerAccessToken));
 
 		try {
 			CustomOAuth2User customOAuth2User = customOAuth2UserService.loadUser(userRequest);
@@ -207,15 +160,7 @@ public class SocialNativeAuthService {
 		log.debug("Verifying OIDC id token and loading user. registrationId={}",
 			clientRegistration.getRegistrationId());
 
-		String normalizedIdToken = normalizeAccessToken(providerIdToken);
-		if (!StringUtils.hasText(normalizedIdToken)) {
-			log.warn("OIDC provider requested without id token. registrationId={}",
-				clientRegistration.getRegistrationId());
-			throw AuthErrorCode.INVALID_TOKEN.toBaseException();
-		}
-
-		Jwt jwt = decodeOidcIdToken(clientRegistration, normalizedIdToken);
-		validateOidcClaims(clientRegistration, jwt);
+		Jwt jwt = oAuthAttributesResolver.decodeAndValidateOidcIdToken(clientRegistration, providerIdToken);
 
 		try {
 			return customOAuth2UserService.loadUserFromOidcClaims(socialType.registrationId(), jwt.getClaims());
@@ -228,97 +173,6 @@ public class SocialNativeAuthService {
 				clientRegistration.getRegistrationId(), e.getClass().getSimpleName());
 			throw AuthErrorCode.INVALID_TOKEN.toBaseException();
 		}
-	}
-
-	/**
-	 * OIDC id token을 검증하고 {@link OAuthAttributes}를 반환합니다. 계정 생성/연동은 수행하지 않습니다.
-	 */
-	private OAuthAttributes extractOidcAttributes(ClientRegistration clientRegistration, String registrationId,
-		String idToken) {
-		String normalizedIdToken = normalizeAccessToken(idToken);
-		if (!StringUtils.hasText(normalizedIdToken)) {
-			throw AuthErrorCode.INVALID_TOKEN.toBaseException();
-		}
-		Jwt jwt = decodeOidcIdToken(clientRegistration, normalizedIdToken);
-		validateOidcClaims(clientRegistration, jwt);
-		return OAuthAttributes.of(registrationId, "sub", jwt.getClaims());
-	}
-
-	/**
-	 * OAuth2 access token으로 provider user-info API를 호출하고 {@link OAuthAttributes}를 반환합니다.
-	 * 계정 생성/연동은 수행하지 않습니다.
-	 */
-	private OAuthAttributes extractOAuth2Attributes(ClientRegistration clientRegistration, String registrationId,
-		String accessToken) {
-		OAuth2UserRequest userRequest = new OAuth2UserRequest(clientRegistration,
-			toProviderAccessToken(accessToken));
-		try {
-			// 카카오 user-info API에 HTTP 요청을 보내서 액세스 토큰 유효성 검증
-			DefaultOAuth2UserService delegate = new DefaultOAuth2UserService();
-			OAuth2User oAuth2User = delegate.loadUser(userRequest);
-			String userNameAttributeName = clientRegistration.getProviderDetails()
-				.getUserInfoEndpoint().getUserNameAttributeName();
-			return OAuthAttributes.of(registrationId, userNameAttributeName, oAuth2User.getAttributes());
-		} catch (RuntimeException e) {
-			log.warn("OAuth2 attributes extraction failed. provider={}, exceptionType={}, message={}",
-				registrationId, e.getClass().getSimpleName(), e.getMessage());
-			throw AuthErrorCode.INVALID_TOKEN.toBaseException();
-		}
-	}
-
-	/**
-	 * provider별 OIDC 설정으로 id token을 디코딩합니다.
-	 * 디코딩 과정에서 서명 및 기본 JWT 검증이 수행됩니다.
-	 */
-	private Jwt decodeOidcIdToken(ClientRegistration clientRegistration, String idToken) {
-		try {
-			JwtDecoder decoder = oidcIdTokenDecoderFactory.createDecoder(clientRegistration);
-			return decoder.decode(idToken);
-		} catch (RuntimeException e) {
-			log.warn("OIDC id token decode failed. provider={}, exceptionType={}",
-				clientRegistration.getRegistrationId(), e.getClass().getSimpleName());
-			throw AuthErrorCode.INVALID_TOKEN.toBaseException();
-		}
-	}
-
-	/**
-	 * OIDC id token의 필수 claim을 검증합니다.
-	 * <ul>
-	 * <li>sub: 소셜 사용자 식별자 존재 여부</li>
-	 * <li>aud: 현재 client_id 포함 여부</li>
-	 * <li>iss: provider별 expected issuer 일치 여부</li>
-	 * </ul>
-	 */
-	private void validateOidcClaims(ClientRegistration clientRegistration, Jwt jwt) {
-		String subject = jwt.getClaimAsString(OIDC_SUB_CLAIM);
-		if (!StringUtils.hasText(subject)) {
-			throw AuthErrorCode.INVALID_SOCIAL_IDENTIFIER.toBaseException();
-		}
-
-		List<String> audience = jwt.getAudience();
-		if (!isAudienceAllowed(clientRegistration, audience)) {
-			throw AuthErrorCode.INVALID_TOKEN.toBaseException();
-		}
-
-		String expectedIssuer = resolveExpectedIssuer(clientRegistration);
-		if (StringUtils.hasText(expectedIssuer)
-			&& (jwt.getIssuer() == null || !expectedIssuer.equals(jwt.getIssuer().toString()))) {
-			throw AuthErrorCode.INVALID_TOKEN.toBaseException();
-		}
-	}
-
-	private boolean isAudienceAllowed(ClientRegistration clientRegistration, List<String> tokenAudience) {
-		if (tokenAudience == null || tokenAudience.isEmpty()) {
-			return false;
-		}
-
-		String clientId = clientRegistration.getClientId();
-		for (String aud : tokenAudience) {
-			if (StringUtils.hasText(aud) && aud.equals(clientId)) {
-				return true;
-			}
-		}
-		return false;
 	}
 
 	private String resolveRegistrationId(String provider, String platform) {
@@ -348,39 +202,6 @@ public class SocialNativeAuthService {
 			return null;
 		}
 		return raw.trim().toLowerCase(Locale.ROOT);
-	}
-
-	/**
-	 * issuer 검증 기준값을 반환합니다.
-	 * provider 설정의 issuerUri를 우선 사용하고, 없으면 provider별 기본값으로 보완합니다.
-	 */
-	private String resolveExpectedIssuer(ClientRegistration clientRegistration) {
-		String issuerUri = clientRegistration.getProviderDetails().getIssuerUri();
-		if (StringUtils.hasText(issuerUri)) {
-			return issuerUri;
-		}
-
-		String registrationId = clientRegistration.getRegistrationId();
-		if (registrationId != null && registrationId.toLowerCase(Locale.ROOT).startsWith("apple")) {
-			return APPLE_ISSUER;
-		}
-
-		if (SocialType.GOOGLE.matchesRegistrationId(registrationId)) {
-			return GOOGLE_ISSUER;
-		}
-
-		return null;
-	}
-
-	private OAuth2AccessToken toProviderAccessToken(String accessToken) {
-		String normalizedAccessToken = normalizeAccessToken(accessToken);
-		if (!StringUtils.hasText(normalizedAccessToken)) {
-			throw AuthErrorCode.INVALID_TOKEN.toBaseException();
-		}
-
-		Instant issuedAt = Instant.now();
-		Instant expiresAt = issuedAt.plusSeconds(PROVIDER_TOKEN_DEFAULT_TTL_SECONDS);
-		return new OAuth2AccessToken(OAuth2AccessToken.TokenType.BEARER, normalizedAccessToken, issuedAt, expiresAt);
 	}
 
 	/**
@@ -418,27 +239,6 @@ public class SocialNativeAuthService {
 			return;
 		}
 		socialAccountOauthRefreshStore.saveEncryptedRefreshToken(user.getId(), socialType, refreshToken);
-	}
-
-	private boolean isOidcProvider(ClientRegistration clientRegistration) {
-		return clientRegistration.getScopes().contains("openid");
-	}
-
-	/**
-	 * access token 입력값을 정규화합니다.
-	 * Bearer 접두어가 포함된 경우 제거합니다.
-	 */
-	private String normalizeAccessToken(String accessToken) {
-		if (!StringUtils.hasText(accessToken)) {
-			return accessToken;
-		}
-
-		String trimmed = accessToken.trim();
-		if (trimmed.regionMatches(true, 0, "Bearer ", 0, 7)) {
-			return trimmed.substring(7).trim();
-		}
-
-		return trimmed;
 	}
 
 }
